# docs/api/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "ReplayBody": {
        "additionalProperties": false,
        "properties": {
          "oem": {
            "maxLength": 32,
            "pattern": "^[a-z0-9_]+$",
            "title": "Oem",
            "type": "string"
          }
        },
        "required": [
          "oem"
        ],
        "title": "ReplayBody",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "OAuth2PasswordBearer": {
        "flows": {
          "password": {
            "scopes": {},
            "tokenUrl": "/api/v1/auth/token"
          }
        },
        "type": "oauth2"
      }
    }
  },
  "info": {
    "description": "\nRosetta translates vehicle telemetry from many car makers into one canonical event.\n\n* **Authentication** OAuth2 bearer tokens (JWT). Get one from `POST /api/v1/auth/token`.\n* **Errors** RFC 9457 problem details (`application/problem+json`).\n* **Pagination** keyset cursors: pass `next_cursor` back as `cursor`. No offsets.\n* **Rate limits** per caller. A 429 carries `Retry-After`.\n* **Versioning** the path carries the major version. Additive changes do not bump it.\n",
    "title": "Rosetta API",
    "version": "1.0.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/dlq/groups": {
      "get": {
        "operationId": "groups_api_v1_dlq_groups_get",
        "parameters": [
          {
            "in": "query",
            "name": "oem",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 32,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Oem"
            }
          },
          {
            "in": "query",
            "name": "open_only",
            "required": false,
            "schema": {
              "default": false,
              "title": "Open Only",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Dead letters grouped by source, reason, field and payload shape",
        "tags": [
          "dead letters"
        ]
      }
    },
    "/api/v1/dlq/groups/{group_id}/samples": {
      "get": {
        "operationId": "samples_api_v1_dlq_groups__group_id__samples_get",
        "parameters": [
          {
            "in": "path",
            "name": "group_id",
            "required": true,
            "schema": {
              "title": "Group Id",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 10,
              "maximum": 50,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Sample payloads of one group",
        "tags": [
          "dead letters"
        ]
      }
    },
    "/api/v1/dlq/replay": {
      "get": {
        "operationId": "replay_jobs_api_v1_dlq_replay_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Replay jobs, newest first",
        "tags": [
          "dead letters"
        ]
      },
      "post": {
        "operationId": "replay_api_v1_dlq_replay_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ReplayBody"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "OAuth2PasswordBearer": []
          }
        ],
        "summary": "Send parked messages of one source through the pipeline again",
        "tags": [
          "dead letters"
        ]
      }
    },
...
  }
}